Dealing with duplicates
Dealing with duplicated articles in the collection.

Option 1: Make the url field a unique index, then load all unique
indexes into memory to filter out duplicates before scraping and
inserting them. This saves overhead on mongo since its an index, and on
R since we’ll have to scrape fewer articles, but it means loading the
distinct url into memory

Option 2: Make url field a non-unique index, allow duplicate records to
be inserted, then identify them and delete them immediately. This saves
us memory but requires mongo aggregation which can be expensive.

Upserting is not an option since each scraped record would have to be
upserted individually and that’s too large a number.

For now, we’ll go with option 1 and experiment to see which one hinders
performance the least.